backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-cloud-storage
python-multipart
openai
google-generativeai
